wikidata_discover/requirements.txt
openai
python-dotenv
rich
pandas
questionary
SPARQLWrapper
requests
